# Conflicts:
#	backend/src/main/java/com/overseer/service/impl/UserServiceImpl.java
#	frontend/src/app/pages/request-table/request-table.component.html
#	frontend/src/app/service/request.service.ts

+++ backend/src/main/java/com/overseer/dao/CrudDao.java
@@ -56,5 +56,5 @@ public interface CrudDao<T, ID extends Serializable> {
      */
     List<T> fetchPage(int pageSize, int pageNumber);
 
-    int count();
+    Long count();
 }
